+++ be/src/main/java/codesquad/issuetracker/controller/MemberController.java
@@ -1,6 +1,6 @@
 package codesquad.issuetracker.controller;
 
-import codesquad.issuetracker.dto.MemberDtoList;
+import codesquad.issuetracker.dto.member.MemberDtoList;
 import codesquad.issuetracker.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
