+++ be/src/main/java/codesquad/issuetracker/service/MemberService.java
@@ -1,6 +1,6 @@
 package codesquad.issuetracker.service;
 
-import codesquad.issuetracker.dto.MemberDtoList;
+import codesquad.issuetracker.dto.member.MemberDtoList;
 import codesquad.issuetracker.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
